+++ lightning/get_channel.js
@@ -30,7 +30,7 @@ const edgeNotFoundErrorMessage = 'edge not found';
     }]
     transaction_id: <Transaction Id Hex String>
     transaction_vout: <Transaction Output Index Number>
-    updated_at: <Channel Last Updated At ISO 8601 Date String>
+    [updated_at]: <Channel Last Updated At ISO 8601 Date String>
   }
 */
 module.exports = ({id, lnd}, cbk) => {