+++ lightning/index.js
@@ -41,6 +41,7 @@ const subscribeToBackups = require('./subscribe_to_backups');
 const subscribeToChannels = require('./subscribe_to_channels');
 const subscribeToGraph = require('./subscribe_to_graph');
 const subscribeToInvoices = require('./subscribe_to_invoices');
+const subscribeToOpenRequests = require('./subscribe_to_open_requests');
 const subscribeToTransactions = require('./subscribe_to_transactions');
 const updateRoutingFees = require('./update_routing_fees');
 const verifyBackup = require('./verify_backup');
@@ -91,6 +92,7 @@ module.exports = {
   subscribeToChannels,
   subscribeToGraph,
   subscribeToInvoices,
+  subscribeToOpenRequests,
   subscribeToTransactions,
   updateRoutingFees,
   verifyBackup,