+++ test/integration/test_subscribe_to_backups.js
@@ -1,3 +1,4 @@
+const asyncRetry = require('async/retry');
 const {test} = require('tap');
 
 const {createCluster} = require('./../macros');
@@ -13,6 +14,8 @@ const channelCapacityTokens = 1e6;
 const confirmationCount = 20;
 const defaultFee = 1e3;
 const giftTokens = 1e5;
+const interval = retryCount => 10 * Math.pow(2, retryCount);
+const times = 20;
 
 // Subscribing to channel backups should trigger backup notifications
 test(`Subscribe to backups`, async ({end, equal}) => {
@@ -21,39 +24,14 @@ test(`Subscribe to backups`, async ({end, equal}) => {
   const {lnd} = cluster.control;
 
   let channelOpen;
+  const got = {};
   const sub = subscribeToBackups({lnd: cluster.control.lnd});
 
   sub.on('error', () => {});
 
-  sub.on('backup', async ({backup, channels}) => {
-    const [channel] = channels;
-
-    const multiVerification = await verifyBackups({
-      backup,
-      lnd,
-      channels: [{
-        transaction_id: channelOpen.transaction_id,
-        transaction_vout: channelOpen.transaction_vout,
-      }],
-    });
-
-    equal(multiVerification.is_valid, true, 'Multiple backups are valid');
-
-    const singleVerification = await verifyBackup({
-      lnd,
-      backup: channel.backup,
-      transaction_id: channelOpen.transaction_id,
-      transaction_vout: channelOpen.transaction_vout,
-    });
-
-    equal(singleVerification.is_valid, true, 'Single backup is valid');
-
-    // Give blocks some time to be generated
-    await delay(3000);
-
-    await cluster.kill({});
-
-    return end();
+  sub.on('backup', ({backup, channels}) => {
+    got.backup = backup;
+    return got.channels = channels;
   });
 
   channelOpen = await openChannel({
@@ -61,11 +39,45 @@ test(`Subscribe to backups`, async ({end, equal}) => {
     chain_fee_tokens_per_vbyte: defaultFee,
     give_tokens: giftTokens,
     local_tokens: channelCapacityTokens,
-    partner_public_key: (await getWalletInfo({lnd})).public_key,
-    socket: `${cluster.control.listen_ip}:${cluster.control.listen_port}`,
+    partner_public_key: cluster.control.public_key,
+    socket: cluster.control.socket,
   });
 
-  await cluster.generate({count: confirmationCount, node: cluster.target});
+  // Wait for generation to be over
+  await asyncRetry({interval, times}, async () => {
+    // Generate to confirm the tx
+    await cluster.generate({count: 1, node: cluster.control});
 
-  return;
+    if (!got.channels) {
+      throw new Error('ExpectedBackupWithChannelsData');
+    }
+
+    return;
+  });
+
+  const [channel] = got.channels;
+
+  const multiVerification = await verifyBackups({
+    lnd,
+    backup: got.backup,
+    channels: [{
+      transaction_id: channelOpen.transaction_id,
+      transaction_vout: channelOpen.transaction_vout,
+    }],
+  });
+
+  equal(multiVerification.is_valid, true, 'Multiple backups are valid');
+
+  const singleVerification = await verifyBackup({
+    lnd,
+    backup: channel.backup,
+    transaction_id: channelOpen.transaction_id,
+    transaction_vout: channelOpen.transaction_vout,
+  });
+
+  equal(singleVerification.is_valid, true, 'Single backup is valid');
+
+  await cluster.kill({});
+
+  return end();
 });