+++ test/routerrpc-integration/test_pay_via_routes.js
@@ -16,6 +16,7 @@ const {hopsFromChannels} = require('./../../routing');
 const {openChannel} = require('./../../');
 const {payViaRoutes} = require('./../../');
 const {routeFromChannels} = require('./../../routing');
+const {setupChannel} = require('./../macros');
 const {waitForChannel} = require('./../macros');
 const {waitForPendingChannel} = require('./../macros');
 
@@ -37,24 +38,10 @@ test(`Pay via routes`, async ({deepIs, end, equal}) => {
   const remoteLnd = cluster.remote.lnd;
   const targetPubKey = cluster.target_node_public_key;
 
-  const controlToTargetChannel = await openChannel({
+  const controlToTargetChan = await setupChannel({
     lnd,
-    chain_fee_tokens_per_vbyte: defaultFee,
-    local_tokens: channelCapacityTokens,
-    partner_public_key: cluster.target_node_public_key,
-    socket: `${cluster.target.listen_ip}:${cluster.target.listen_port}`,
-  });
-
-  await waitForPendingChannel({
-    lnd,
-    id: controlToTargetChannel.transaction_id,
-  });
-
-  await cluster.generate({count: confirmationCount, node: cluster.control});
-
-  const controlToTargetChan = await waitForChannel({
-    lnd,
-    id: controlToTargetChannel.transaction_id,
+    generate: cluster.generate,
+    to: cluster.target,
   });
 
   const [channel] = (await getChannels({lnd})).channels;